requirements.txt
asgiref==3.7.2
click==8.1.7
docx2txt==0.8
docxtpl==0.16.7
groq==0.4.2
Jinja2==3.1.3
langchain==0.1.12
langchain-core==0.1.32
langchain-groq==0.0.2
python-docx==1.1.0
python-dotenv==1.0.1
PyPDF2==3.0.1
streamlit==1.32.1